+++ src/type-extensions.ts
@@ -28,3 +28,10 @@ declare global {
     }
   }
 }
+
+// To prevent an error where JSON.stringify() doesn't know how to serialize a BigInt,
+// we need to add a toJSON method to BigInt's prototype.
+// Refer to the MDN documentation for more details: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt#use_within_json
+(BigInt.prototype as any).toJSON = function () {
+  return this.toString();
+};